remove unused concurrency test
